# server/openapi.yaml
openapi: 3.1.0
info:
  title: CA Connection Gateway API
  version: 0.4.1
  description: >
    Production API for Android gateway ingestion, authenticated SMS and
    notification retrieval, remote outbound SMS commands, device management,
    pairing, and one-time OTP claims. Sensitive request and response bodies
    must never be logged.
servers:
  - url: https://gateway.example.com
paths:
...
  /v1/notification-channels:
    get:
      summary: List notification channels without exposing provider credentials
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Configured channels and delivery summaries
        "401":
          description: API authentication failed
        "403":
          description: Token lacks notifications:manage or global device access
  /v1/notification-channels/{channelId}:
    put:
      summary: Update one channel's enablement, privacy mode, and event subscriptions
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: channelId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enabled, contentMode, eventTypes]
              properties:
                enabled:
                  type: boolean
                contentMode:
                  type: string
                  enum: [REDACTED, FULL]
                eventTypes:
                  type: array
                  minItems: 1
                  uniqueItems: true
                  items:
                    type: string
                    enum:
                      - sms.received
                      - call.ringing
                      - call.missed
                      - call.ended
                      - notification.received
      responses:
        "200":
          description: Channel updated
        "400":
          description: Invalid channel settings
        "401":
          description: API authentication failed
        "403":
          description: Token lacks notifications:manage or global device access
        "404":
          description: Channel is not configured
  /v1/notification-channels/{channelId}/test:
    post:
      summary: Queue a non-sensitive test for one notification channel
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: channelId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "202":
          description: Test notification queued
        "400":
          description: Request body must be empty
        "401":
          description: API authentication failed
        "403":
          description: Token lacks notifications:manage or global device access
        "404":
          description: Channel is not configured
        "409":
          description: Notification dispatcher is unavailable
  /v1/notification-settings:
    get:
      summary: Read Feishu delivery state without exposing Webhook credentials
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification settings and delivery summary
        "401":
          description: API authentication failed
        "403":
          description: Token lacks notifications:manage or global device access
    put:
      summary: Enable or disable Feishu delivery and select content privacy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enabled, contentMode]
              properties:
                enabled:
                  type: boolean
                contentMode:
                  type: string
                  enum: [REDACTED, FULL]
      responses:
        "200":
          description: Settings updated
        "400":
          description: Invalid settings
        "401":
          description: API authentication failed
        "403":
          description: Token lacks notifications:manage or global device access
        "409":
          description: Feishu Webhook is not configured
  /v1/notification-settings/test:
    post:
      summary: Queue a non-sensitive Feishu test notification
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "202":
          description: Test notification queued
        "400":
          description: Request body must be empty
        "401":
          description: API authentication failed
        "403":
          description: Token lacks notifications:manage or global device access
        "409":
          description: Feishu Webhook is not configured
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...